+++ src/Agent/NewRelic.Api.Agent/NoOpSpan.cs
@@ -9,5 +9,10 @@ namespace NewRelic.Api.Agent
         {
             return this;
         }
+
+        public ISpan SetName(string name)
+        {
+            return this;
+        }
     }
 }