+++ src/Agent/NewRelic/Agent/Core/Agent.cs
@@ -15,6 +15,7 @@ using NewRelic.Agent.Core.Segments;
 using NewRelic.Agent.Core.Transactions;
 using NewRelic.Agent.Core.Transformers.TransactionTransformer;
 using NewRelic.Agent.Core.Utilities;
+using NewRelic.Agent.Core.Utils;
 using NewRelic.Agent.Core.WireModels;
 using NewRelic.Agent.Core.Wrapper.AgentWrapperApi.Builders;
 using NewRelic.Agent.Core.Wrapper.AgentWrapperApi.CrossApplicationTracing;
@@ -36,6 +37,7 @@ namespace NewRelic.Agent.Core
     public class Agent : IAgent // any changes to api, update the interface in extensions and re-import, then implement in legacy api as NotImplementedException
     {
         public const int QueryParameterMaxStringLength = 256;
+        private const int LogExceptionStackLimit = 300;
         internal static Agent Instance;
         private static readonly ITransaction _noOpTransaction = new NoOpTransaction();
 
@@ -406,7 +408,7 @@ namespace NewRelic.Agent.Core
             _agentHealthReporter.ReportSupportabilityCountMetric(metricName, count);
         }
 
-        public void RecordLogMessage(string frameworkName, object logEvent, Func<object, DateTime> getTimestamp, Func<object, object> getLevel, Func<object, string> getLogMessage, string spanId, string traceId)
+        public void RecordLogMessage(string frameworkName, object logEvent, Func<object, DateTime> getTimestamp, Func<object, object> getLevel, Func<object, string> getLogMessage, Func<object, Exception> getLogException, string spanId, string traceId)
         {
             _agentHealthReporter.ReportLogForwardingFramework(frameworkName);
 
@@ -427,30 +429,48 @@ namespace NewRelic.Agent.Core
             if (_configurationService.Configuration.LogEventCollectorEnabled)
             {
                 var logMessage = getLogMessage(logEvent);
-                if (string.IsNullOrWhiteSpace(logMessage))
+                var logException = getLogException(logEvent);
+
+                // exit quickly if the message and exception are missing
+                if (string.IsNullOrWhiteSpace(logMessage) && logException is null)
                 {
                     return;
                 }
+
                 var timestamp = getTimestamp(logEvent).ToUnixTimeMilliseconds();
+
+                LogEventWireModel logEventWireModel = null;
+                if (logException != null)
+                {
+                    logEventWireModel = new LogEventWireModel(timestamp, logMessage, normalizedLevel,
+                        StackTraces.ScrubAndTruncate(logException, LogExceptionStackLimit), logException.Message, logException.GetType().ToString(),
+                        spanId, traceId);
+                }
+                else
+                {
+                    logEventWireModel = new LogEventWireModel(timestamp, logMessage, normalizedLevel, spanId, traceId);
+                }
 
                 var transaction = _transactionService.GetCurrentInternalTransaction();
                 if (transaction != null && transaction.IsValid)
                 {
                     // use transaction batching for messages in transactions
-                    if (!transaction.AddLogEvent(new LogEventWireModel(timestamp, logMessage, normalizedLevel, spanId, traceId)))
+                    if (!transaction.AddLogEvent(logEventWireModel))
                     {
                         // AddLogEvent returns false in the case that logs have already been harvested by transaction transform.
                         // Fall back to collecting the log based on the information we have. Since the transaction was finalized,
                         // the Priority should be correct.
-                        _logEventAggregator.Collect(new LogEventWireModel(timestamp, logMessage, normalizedLevel, spanId,
-                            traceId, transaction.Priority));
+                        logEventWireModel.Priority = transaction.Priority;
+                        _logEventAggregator.Collect(logEventWireModel);
+
                     }
+
                     return;
                 }
 
                 // non-transaction messages with proper sanitized priority value
-                _logEventAggregator.Collect(new LogEventWireModel(timestamp,
-                    logMessage, normalizedLevel, spanId, traceId, _transactionService.CreatePriority()));
+                logEventWireModel.Priority = _transactionService.CreatePriority();
+                _logEventAggregator.Collect(logEventWireModel);
             }
         }