+++ src/Agent/NewRelic/Agent/Core/Configuration/DefaultConfiguration.cs
@@ -893,6 +893,7 @@ namespace NewRelic.Agent.Core.Configuration
 
         public int? SamplingTarget => _serverConfiguration.SamplingTarget;
 
+        // Faster Event Harvest configuration rules apply here, which is why ServerOverrides takes precedence over EnvironmentOverrides
         public int SpanEventsMaxSamplesStored => ServerOverrides(_serverConfiguration.SpanEventHarvestConfig?.HarvestLimit,
            EnvironmentOverrides(_localConfiguration.spanEvents.maximumSamplesStored, "NEW_RELIC_SPAN_EVENTS_MAX_SAMPLES_STORED").GetValueOrDefault());
 
@@ -1124,6 +1125,8 @@ namespace NewRelic.Agent.Core.Configuration
         {
             get
             {
+                // Faster Event Harvest configuration rules apply here, if/when we add an environment variable for this property we need to make sure that
+                // ServerOverrides takes precedence over EnvironmentOverrides
                 return ServerOverrides(_serverConfiguration.EventHarvestConfig?.ErrorEventHarvestLimit(), _localConfiguration.errorCollector.maxEventSamplesStored);
             }
         }
@@ -1528,10 +1531,9 @@ namespace NewRelic.Agent.Core.Configuration
         {
             get
             {
-                return (int)EnvironmentOverrides(
-                    ServerOverrides(_serverConfiguration.EventHarvestConfig?.CustomEventHarvestLimit(),
-                        _localConfiguration.customEvents.maximumSamplesStored),
-                    "MAX_EVENT_SAMPLES_STORED");
+                // Faster Event Harvest configuration rules apply here, which is why ServerOverrides takes precedence over EnvironmentOverrides
+                var maxValue = _localConfiguration.customEvents.maximumSamplesStored;
+                return ServerOverrides(_serverConfiguration.EventHarvestConfig?.CustomEventHarvestLimit(), (int)EnvironmentOverrides(maxValue, "MAX_EVENT_SAMPLES_STORED"));
             }
         }
 
@@ -1599,10 +1601,10 @@ namespace NewRelic.Agent.Core.Configuration
         {
             get
             {
+                // Faster Event Harvest configuration rules apply here, which is why ServerOverrides takes precedence over EnvironmentOverrides
                 var maxValue = _localConfiguration.transactionEvents.maximumSamplesStored;
-                return (int)EnvironmentOverrides(
-                    ServerOverrides(_serverConfiguration.EventHarvestConfig?.TransactionEventHarvestLimit(), maxValue),
-                    "MAX_TRANSACTION_SAMPLES_STORED");
+
+                return ServerOverrides(_serverConfiguration.EventHarvestConfig?.TransactionEventHarvestLimit(), (int)EnvironmentOverrides(maxValue, "MAX_TRANSACTION_SAMPLES_STORED"));
             }
         }