+++ src/Agent/NewRelic/Agent/Extensions/Providers/Wrapper/SerilogLogging/NewRelicSerilogSink.cs
@@ -34,12 +34,12 @@ namespace NewRelic.Providers.Wrapper.SerilogLogging
 
             Func<object, DateTime> getDateTimeFunc = l => logEvent.Timestamp.UtcDateTime;
 
+            Func<object, Exception> getLogExceptionFunc = l => logEvent.Exception;
+
             Func<object, string> getMessageFunc = l => logEvent.RenderMessage();
 
-            // This will either add the log message to the transaction or directly to the aggregator
-
             var xapi = _agent.GetExperimentalApi();
-            xapi.RecordLogMessage("serilog", logEvent, getDateTimeFunc, getLevelFunc, getMessageFunc, _agent.TraceMetadata.SpanId, _agent.TraceMetadata.TraceId);
+            xapi.RecordLogMessage("serilog", logEvent, getDateTimeFunc, getLevelFunc, getMessageFunc, getLogExceptionFunc, _agent.TraceMetadata.SpanId, _agent.TraceMetadata.TraceId);
         }
     }
 }