+++ tests/Agent/IntegrationTests/SharedApplications/Common/MultiFunctionApplicationHelpers/NetStandardLibraries/LogInstrumentation/MicrosoftLoggingLoggingAdapter.cs
@@ -4,6 +4,7 @@
 
 #if NETCOREAPP2_1_OR_GREATER || NET48_OR_GREATER
 
+using System;
 using Microsoft.Extensions.Logging;
 using Serilog;
 using Serilog.Formatting.Json;
@@ -33,14 +34,24 @@ namespace MultiFunctionApplicationHelpers.NetStandardLibraries.LogInstrumentatio
             logger.LogWarning(message);
         }
 
-        public void Error(string message)
+        public void Error(Exception exception)
         {
-            logger.LogError(message);
+            logger.LogError(exception, exception.Message);
+        }
+
+        public void ErrorNoMessage(Exception exception)
+        {
+            logger.LogError(exception, string.Empty);
         }
 
         public void Fatal(string message)
         {
             logger.LogCritical(message);
+        }
+
+        public void NoMessage()
+        {
+            logger.LogTrace(string.Empty);
         }
 
         public void Configure()